chore: move NodeRecord type
