+++ Cargo.toml
@@ -338,6 +338,7 @@ smallvec = "1"
 dyn-clone = "1.0.17"
 sha2 = { version = "0.10", default-features = false }
 paste = "1.0"
+url = "2.3"
 
 # proc-macros
 proc-macro2 = "1.0"
