+++ crates/net/types/Cargo.toml
@@ -12,17 +12,22 @@ description = "Network types and utils"
 workspace = true
 
 [dependencies]
-# reth
-reth-rpc-types.workspace = true
-
-alloy-primitives.workspace = true
 
 # eth
+alloy-primitives = { workspace = true, features = ["rlp"] }
+alloy-rlp = { workspace = true, features = ["derive"] }
 enr.workspace = true
 
 # crypto
-secp256k1 = { workspace = true, features = ["global-context", "recovery", "rand"] }
+secp256k1.workspace = true
 
 # misc
 serde_with.workspace = true
+thiserror.workspace = true
+url.workspace = true
 
+[dev-dependencies]
+alloy-primitives = { workspace = true, features = ["rand"] }
+rand.workspace = true
+secp256k1 = { workspace = true, features = ["rand"] }
+serde_json.workspace = true
