+++ crates/primitives/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 # reth
 reth-codecs.workspace = true
 reth-ethereum-forks.workspace = true
+reth-network-types.workspace = true
 reth-rpc-types.workspace = true
 revm.workspace = true
 revm-primitives = { workspace = true, features = ["serde"] }
